requirements/base.txt
Django==5.0.6
django-environ==0.11.2
psycopg2==2.9.9
djangorestframework==3.14.0
django-easy-audit==1.3.6
django-guardian==2.4.0

whitenoise==6.5.0
pillow==10.3.0

django-filter==23.5
django-extensions==3.2.3
django-cors-headers==4.3.1
django-storages==1.14.2
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10

drf-spectacular==0.27.2

channels==4.1.0
daphne==4.1.2
uwsgi==2.0.25.1
gunicorn==22.0.0
setuptools==70.0.0
